Moving component from overview utils to components folder
