+++ src/views/clusteroverview/ClusterOverviewPage.tsx
@@ -1,28 +1,18 @@
 import * as React from 'react';
 import { Helmet } from 'react-helmet';
 
-import { FLAG_CONSOLE_CLI_DOWNLOAD } from '@kubevirt-utils/flags/consts';
 import { useIsAdmin } from '@kubevirt-utils/hooks/useIsAdmin';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
-import { HorizontalNav, NavPage, useFlag } from '@openshift-console/dynamic-plugin-sdk';
+import { HorizontalNav, NavPage } from '@openshift-console/dynamic-plugin-sdk';
 
+import ClusterOverviewPageHeader from './Header/ClusterOverviewPageHeader';
 import MonitoringTab from './MonitoringTab/MonitoringTab';
 import OverviewTab from './OverviewTab/OverviewTab';
 import SettingsTab from './SettingsTab/SettingsTab';
-import { KUBEVIRT_QUICK_START_USER_SETTINGS_KEY } from './utils/constants';
-import PageHeader from './utils/PageHeader';
-import RestoreGettingStartedButton from './utils/RestoreGettingStartedButton';
-import VirtctlPopup from './utils/VirtctlPopup';
 
 const ClusterOverviewPage: React.FC = () => {
   const { t } = useKubevirtTranslation();
-  const title = t('Virtualization');
   const isAdmin = useIsAdmin();
-  const hasConsoleTools = useFlag(FLAG_CONSOLE_CLI_DOWNLOAD);
-
-  const badge = (
-    <RestoreGettingStartedButton userSettingsKey={KUBEVIRT_QUICK_START_USER_SETTINGS_KEY} />
-  );
 
   const overviewTabs: NavPage[] = [
     {
@@ -45,11 +35,9 @@ const ClusterOverviewPage: React.FC = () => {
   return (
     <>
       <Helmet>
-        <title>{title}</title>
+        <title>{t('Virtualization')}</title>
       </Helmet>
-      <PageHeader title={title} badge={badge}>
-        {hasConsoleTools && <VirtctlPopup />}
-      </PageHeader>
+      <ClusterOverviewPageHeader />
       {isAdmin ? <HorizontalNav pages={overviewTabs} /> : <OverviewTab />}
     </>
   );
